1.0.1.0 Minor enhancements, controller support, keybinds and Linux executable
I made some new particles, added controller support and made a level select screen.

The game is now available on Linux.

The game is now using compiled native code instead of interpreted code, which is recommended for logic-heavy games (Shadow Dash is NOT logic-heavy, but I am using it anyways).